hide secret key and fix email payment receipts
