+++ .gitignore
@@ -44,3 +44,4 @@ imports/plugins/custom/*
 !imports/plugins/custom/.gitkeep
 
 .reaction/config.json
+.env
