+++ imports/plugins/included/payments-paystack/client/index.js
@@ -1,3 +1,3 @@
 import "./checkout/paystack";
-import "./settings/templates/paystack";
+import "./settings/paystack";
